Refactor
Moved out functions from SageIdeator.tsx by concern - Namely image generation, idea generation, and importation of stickies
